fix(NCP): Show pledge page if pledged collective
