+++ lib/graphql/queries.js
@@ -336,6 +336,7 @@ export const getCollectiveQuery = gql`
     Collective(slug: $slug) {
       id
       isActive
+      isPledged
       type
       slug
       path
@@ -503,20 +504,6 @@ export const getCollectiveQuery = gql`
               slug
             }
           }
-        }
-      }
-      pledges: orders(status: PENDING) {
-        currency
-        id
-        interval
-        publicMessage
-        status
-        totalAmount
-        fromCollective {
-          name
-          imageUrl
-          slug
-          type
         }
       }
     }
